test unit basic
test basic.

+++ TestesUnitarios/src/main/java/com/calculator/module/interfaces/Calculator.java
@@ -3,43 +3,48 @@ package com.calculator.module.interfaces;
 public interface Calculator {
     /**
      * Realiza a adição de dois números
+     *
      * @param a primeiro operando
      * @param b segundo operando
      * @return resultado da adição
      */
-    double addition(double a, double b);
+    int addition(int a, int b);
 
     /**
      * Realiza a subtração de dois números
+     *
      * @param a primeiro operando
      * @param b segundo operando
      * @return resultado da subtração
      */
-    double subtraction(double a, double b);
+    int subtraction(int a, int b);
 
     /**
      * Realiza a multiplicação de dois números
+     *
      * @param a primeiro operando
      * @param b segundo operando
      * @return resultado da multiplicação
      */
-    double multiplication(double a, double b);
+    int multiplication(int a, int b);
 
     /**
      * Realiza a divisão de dois números
+     *
      * @param a primeiro operando
      * @param b segundo operando
      * @return resultado da divisão
      * @throws ArithmeticException quando o divisor for zero
      */
-    double division(double a, double b) throws ArithmeticException;
+    int division(int a, int b) throws ArithmeticException;
 
     /**
      * Realiza a divisão com resto de dois números
+     *
      * @param a primeiro operando
      * @param b segundo operando
      * @return resto da divisão
      * @throws ArithmeticException quando o divisor for zero
      */
-    double restDivision(double a, double b) throws ArithmeticException;
+    int restDivision(int a, int b) throws ArithmeticException;
 }
