+++ TestesUnitarios/src/main/java/com/calculator/module/operators/Addition.java
@@ -3,11 +3,11 @@ package com.calculator.module.operators;
 public class Addition {
     /**
      * Realiza a adição de dois números
+     *
      * @param a primeiro operando
-     * @param b segundo operando
      * @return resultado da adição
      */
-    public double calculate(double a, double b) {
+    public int calculate(int a, int b) {
         return a + b;
     }
 }
